src: game: add multiple collision volumes per entity
Every entity has collision volumes. This enables us to have multiple
collision points for entity. (e.g. hero's torso and leg have different
collision volumes). This is implemented as 1 collision per entity for
now.

We do not need for StoredEntityAddGrounded() because position on entity
is on ground. We have offset for collisions from center.

1.1f needed for stairs to collide. Otherwise if you go up with stair,
you can not back down because of same z.