src: linux: add continuously rendering
Wayland would not send wl_surface_frame::done event for some time when
you are not focused on that surface. This causes elapsed time from last
render to become high (over 1sec). Game crashed if we give abnormal
delta time. Expected delta time is ~33ms for 30fps.

If elapsed time from last render over a second, rewind that to a frame's
worth delta time for continuous rendering. Meaning give ~33ms regardless
of last render time.

If we are pausing the rendering, use same trick when focus is on surface
again.