src: render_group: embed alignment into bitmap structure
Alignment is used frequently so embed into bitmap structure so that we
no longer have specify at every render call. It does not make sense to
specify alignment for bitmap in render calls. Alignment is in pixels, so
including it in render calls, requires caller to change thinking from
meters to pixels for just alignment.