src: asset: use given memory to dynamically allocate assets
Instead of allocating memory from platform, use given memory arena.

Split given arena into blocks. When allocating set flag to used and if
there is space available split into block. When releasing memory set
flag to unused, so that it shows up in memory search for given size.

This code is not final.