src: asset: fix asset system thread bug
Before we evict assets at end of the frame to make sure they are not
used anymore. (4d12120e67a64edc5433890b03ad89f633fc81f7) But we break
that logic when we try to evict assets when acquiring we low on memory.

Replaced by generation id to track if it in use. This is in draft
status.